+++ kshell/src/main/kotlin/kotlinx/cli/CommandLineException.kt
@@ -1,0 +1,13 @@
+package kotlinx.cli
+
+class CommandLineException(message: String) : RuntimeException(message)
+
+open class StopParsingException : RuntimeException()
+
+class HelpPrintedException : StopParsingException()
+
+class MissingArgumentException : RuntimeException()
+
+fun stopParsing() {
+    throw StopParsingException()
+}