+++ kshell/src/main/kotlin/sparklin/kshell/Plugin.kt
@@ -1,9 +1,0 @@
-package sparklin.kshell
-
-import sparklin.kshell.configuration.Configuration
-
-interface Plugin {
-    fun init(repl: KShell, config: Configuration)
-
-    fun cleanUp()
-}