+++ kshell/src/main/kotlin/sparklin/kshell/plugins/SparkPlugin.kt
@@ -1,8 +1,0 @@
-package sparklin.kshell.plugins
-
-import sparklin.kshell.Plugin
-import org.apache.hadoop.conf.Configuration
-
-interface SparkPlugin : Plugin {
-    fun hadoopConfiguration(): Configuration?
-}