+++ kshell/src/test/kotlin/com/github/khud/sparklin/kshell/plugins/PromptPluginTest.kt
@@ -1,0 +1,12 @@
+package com.github.khud.sparklin.kshell.plugins
+
+import org.junit.Assert.*
+import org.junit.Test
+
+class PromptPluginTest {
+    @Test
+    fun testFormat() {
+        val types = mapOf("s" to { "S" }, "ss" to {"Z"})
+        assertEquals("hello SZ{} %d", PromptPlugin.format("hello %s%{ss}{} %d", types))
+    }
+}